refactored isUserAuthorised
